src/pinging_service/pyproject.toml
[tool.poetry]
name = "pinging-service"
version = "0.1.0"
description = "Pinging service to alive_check app"
authors = ["Jakub Jadczak <jakubjadczak02@aoutlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.1"
icmplib = "^3.0.4"
python-dotenv = "^1.0.1"
pika = "^1.3.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
